Refactored one line text field
