+++ uikit/src/main/java/care/intouch/uikit/ui/utill/StringVO.kt
@@ -1,4 +1,4 @@
-package care.intouch.uikit.ui.util
+package care.intouch.uikit.ui.utill
 
 import androidx.annotation.StringRes
 import androidx.compose.runtime.Composable
